requirements.txt
asgiref==3.3.1
Django==3.1.7
psycopg2-binary
whitenoise
pytz==2021.1
sqlparse==0.4.1
